requirements.txt
telethon
pyyaml
python-dotenv
pre-commit
black
ruff
mypy
types-PyYAML
